Split adapters into their own modules
