+++ saas/rti/adapters/processor_scripts.py
@@ -4,6 +4,7 @@ import subprocess
 
 # TODO: Create cli interface for this script
 
+
 def clone_repo(git_local_path: str, git_spec: dict):
     git_url = git_spec['source']
     commit_id = git_spec['commit_id']
