--- tools/template/descriptor_template.json
@@ -1,18 +1,0 @@
-{
-    "name": "{{name}}",
-    "version": "1",
-    "input": [
-        {
-            "name": "a",
-            "data_type": "integer",
-            "data_format": "json"
-        }
-    ],
-    "output": [
-        {
-            "name": "b",
-            "data_type": "integer",
-            "data_format": "json"
-        }
-    ]
-}
