--- tools/template/folder_structure/dependencies/README.md
@@ -1,1 +1,0 @@
-Any external programs/dependencies or files should be stored here.
