--- tools/template/folder_structure/modules/README.md
@@ -1,1 +1,0 @@
-Any other python modules/scripts other than processor.py should be stored here.
